worker_agent/requirements.txt
pika==1.2.0
psycopg2-binary==2.9.1
prometheus-client==0.11.0
python-dotenv==0.19.0
ollama==0.1.0
colorlog==6.7.0 